homework/10-practice-security/auth/requirements.txt
flask
requests==2.28.1
PyJWT==2.6.0
cryptography==38.0.4
